fastApiProject/pyproject.toml
[tool.poetry]
name = "BirbBerry"
version = "0.1.0"
description = ""
authors = ["Viktorija"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
alembic = "^1.12.1"
sqlmodel="^0.0.8"
asyncpg="^0.29.0"
bcrypt="^4.0.1"
email-validator="^2.1.0.post1"
python-jose="^3.3.0"
python-multipart="^0.0.6"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
